hybrid approach to finding the relevent info..
Since it's not possible to determine at the database level what the
title of the advertising URL will be, this commit takes a hybrid
approach.

The idea is to read the provided XML document, then use the
council_reference value to search the list of advertising documents for
the relevent url.